% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{bootstraps}
\alias{bootstraps}
\title{Yank a column}
\usage{
bootstraps(.data, var)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{var}{Column to extract}
}
\value{
A vector
}
\description{
I pulled myself up by my bootstraps (pull equivalent)
}
\examples{
mtcars |> bootstraps(mpg)
}
