% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{ceasefire}
\alias{ceasefire}
\title{Call a ceasefire}
\usage{
ceasefire(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}

\item{by}{Columns to join by}
}
\value{
A merged data frame
}
\description{
We have to go along to get along, so call a ceasefire (inner join equivalent)
}
