% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{chuck_it}
\alias{chuck_it}
\title{Chuck garbage rows}
\usage{
chuck_it(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Logical predicates to keep rows}
}
\value{
A data frame with yeeted rows removed
}
\description{
Don't need all those rows? Just chuck it in the fuggit bucket (filter equivalent)
}
\examples{
mtcars |> chuck_it(mpg > 20, cyl == 4)
}
