% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{divide_and_conquer}
\alias{divide_and_conquer}
\title{Take the bull by the horns}
\usage{
divide_and_conquer(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to group by}
}
\value{
A grouped data frame
}
\description{
Let's you-me divide and conquer this data, bucko (group_by equivalent)
}
\examples{
mtcars |> divide_and_conquer(cyl) |> the_skinny(avg_mpg = mean(mpg))
}
