% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{fiddle}
\alias{fiddle}
\title{Fiddle with it}
\usage{
fiddle(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs of expressions}
}
\value{
A data frame with new/modified columns
}
\description{
Fiddle with the data until she's where ya want her (mutate equivalent)
}
\examples{
mtcars |> fiddle(mpg_squared = mpg^2, high_hp = hp > 150)
}
