% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{get_act_together}
\alias{get_act_together}
\title{Get your act together}
\usage{
get_act_together(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to sort by}
}
\value{
A sorted data frame
}
\description{
I've been up since 5 AM (arrange equivalent)
}
\examples{
mtcars |> get_act_together(desc(mpg), hp)
}
