% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{kumbaya}
\alias{kumbaya}
\title{Bring everyone together for a kumbaya session}
\usage{
kumbaya(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A merged data frame
}
\description{
Let's just hash this out (full join equivalent)
}
