% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{out_of_the_will}
\alias{out_of_the_will}
\title{Those kids that don't visit?}
\usage{
out_of_the_will(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A data frame with non-matching rows
}
\description{
Out of the will (anti_join equivalent)
}
