% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{paid_my_dues}
\alias{paid_my_dues}
\title{I paid my dues, pal}
\usage{
paid_my_dues(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Optional columns to determine uniqueness}
}
\value{
A data frame with unique rows
}
\description{
I deserve this distinction (distinct equivalent)
}
\examples{
mtcars |> paid_my_dues(cyl)
}
