% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{peacekeeping_mission}
\alias{peacekeeping_mission}
\title{This is a peacekeeping mission}
\usage{
peacekeeping_mission(.data, n = 6)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{n}{Number of rows to keep}
}
\value{
A data frame with first n rows
}
\description{
We won't slice any heads (head/slice_head equivalent)
}
\examples{
mtcars |> peacekeeping_mission(5)
}
