% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{pension}
\alias{pension}
\title{You guys have pensions?}
\usage{
pension(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Variables to count by}
}
\value{
A data frame with counts
}
\description{
Cash your pension out (count equivalent)
}
\examples{
mtcars |> pension(cyl, gear)
}
