% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{pick_your_poison}
\alias{pick_your_poison}
\title{Pick your columns and get it over with}
\usage{
pick_your_poison(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Columns to keep}
}
\value{
A data frame with selected columns
}
\description{
You don’t need those columns, we have columns at home (select equivalent)
}
\examples{
mtcars |> pick_your_poison(mpg, cyl, hp)
}
