% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{rebrand}
\alias{rebrand}
\title{Column name too harsh?}
\usage{
rebrand(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs (new_name = old_name)}
}
\value{
A data frame with renamed columns
}
\description{
We can just rebrand (rename equivalent)
}
\examples{
mtcars |> rebrand(miles_per_gallon = mpg, horse_power = hp)
}
