% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{rubber_stamp}
\alias{rubber_stamp}
\title{The right table rubber stamps}
\usage{
rubber_stamp(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble (main table)}

\item{y}{A data frame or tibble (rubber-stamping table)}

\item{by}{Columns to join by}
}
\value{
A merged data frame
}
\description{
Merge, but only keep rows that are rubber stamped by the right table (right join equivalent)
}
