% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{the_skinny}
\alias{the_skinny}
\title{Get the straight skinny}
\usage{
the_skinny(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Name-value pairs of summary expressions}
}
\value{
A data frame with summary statistics
}
\description{
Summary stats (summarize equivalent)
}
\examples{
mtcars |> the_skinny(avg_mpg = mean(mpg), max_hp = max(hp))
}
