% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{trim_the_fat}
\alias{trim_the_fat}
\title{Times are tough}
\usage{
trim_the_fat(x, y, ...)
}
\arguments{
\item{x}{A data frame or tibble}

\item{y}{A data frame or tibble}
}
\value{
A data frame with matching rows
}
\description{
Trim the fat and remove rows that don't match (semi join equivalent)
}
