% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boomerplyr.R
\name{yoko}
\alias{yoko}
\title{Break up the band}
\usage{
yoko(.data)
}
\arguments{
\item{.data}{A grouped data frame}
}
\value{
An ungrouped data frame
}
\description{
Every data frame has a Yoko (ungroup equivalent)
}
\examples{
mtcars |> divide_and_conquer(cyl) |> yoko()
}
